CADisp/requirements.txt
pygame
PyQt5
pyzmq
requests